requirements.txt
fastapi
uvicorn
python-dotenv
cassandra-driver
email-validator
argon2-cffi
pytest
jinja2
python-multipart
python-jose